Refactor project structure with modular packages
